--- test_drivers/test_graph.py
@@ -1,1 +1,0 @@
-from Food import FoodItem
